requirement.txt
snowflake-connector-python
azure-storage-blob
pandas
matplotlib
seaborn
jupyter